chore: improve Grid exports
